[EditContext] Enable SelectAll in context menu while using edit context
Currently, Document::queryCommandEnabled is used to determine if
selectAll should be enabled in the context menu. This creates a
selectAll EditorCommand with kDOM as EditorCommandSource. However,
for commands triggered by user input, kMenuOrKeyBinding should be
used as EditorCommandSource.
This CL updates the condition to check if selectAll should be
enabled in the context menu when edit context is used for an
element by using kMenuOrKeyBinding as EditorCommandSource.

Bug: 371791303
Change-Id: I333a0bde5f08362abfdb43e557be9d05a61587f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6285786
Reviewed-by: Dan Clark <daniec@microsoft.com>
Commit-Queue: Ashish Kumar <ashishkum@microsoft.com>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1422983}
NOKEYCHECK=True
GitOrigin-RevId: d9cd04be7698ea649e52718e812a7210a05e9e6c