[functions] Add InspectorCSSAgent::CollectReferencedFunctionRules
This CL adds a function that collects all referenced custom functions
(transitively) within some match result. This is intended to be used
during getMatchedStylesForNode, but it's unused outside of tests
for now.

Function references (<dashed-function>s) are tree-scoped references,
which means that they can only be dereferenced in the context
of the originating tree-scope [1]. The code in this CL generally
prepares for handling this correctly, but just uses the Document
for now (pending Issue 394111301).

[1] https://drafts.csswg.org/css-scoping-1/#css-tree-scoped-reference

Bug: 325504770, 392034817, 394111301
Change-Id: I1269f403b96fc5c21de938fa2402dc0958780bdc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6253679
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1423099}
NOKEYCHECK=True
GitOrigin-RevId: 9fa89c8f8bfd46f83916e885887a89bd3d93dee4