Add an img_reference reftest type in canvas WPT test generator
In the process of documenting the test generator, all template
parameters should have a clear purpose that can be explained. The
templates used to generate the test files use the parameter
'img_reference'. This parameter is populated by the test generator,
using the filename of the image generated when using 'cairo_reference'.
Seeing that the variable exists, users of the test generator could be
tempted to use it. We must therefore decide whether we should allow
this or whether we should make this a reserved internal parameter.

Allowing the use of this parameter could actually be useful. Tests
could use it to test against solid color images, like
'/images/green-100x50.png'. This would be much better than how lots of
tests currently specify 'expected: green' in their definition. The
images specified with 'expected:' are not used by the test runner to
check the test result, but 'img_reference' is.

Bug: 393561071
Change-Id: I8c5a6a5ec940a988b49fac61075ef99e6279ab69
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6219869
Commit-Queue: Jean-Philippe Gravel <jpgravel@chromium.org>
Reviewed-by: Andres Ricardo Perez <andresrperez@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1423137}
NOKEYCHECK=True
GitOrigin-RevId: a1d387d865466df763dc9144e082c02c87288626