Support aggregate error reporting in the PrivateAggregationHost
Adds support for the kPrivateAggregationApiErrorReporting feature,
including its ContributeToHistogramOnEvent() method. When the feature is
enabled, delegates triggering those conditional contributions (as well
as general merging and truncation) to the
PrivateAggregationPendingContributions object. Also allows debug keys to
be set on null reports.

When the feature is disabled (which it is by default), this cl should
have no effect.

Bug: 381788013
Change-Id: If10f2236a54ab08f916ed02c7bc92b74a2d1f435
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6207481
Reviewed-by: Dan McArdle <dmcardle@chromium.org>
Commit-Queue: Alex Turner <alexmt@chromium.org>
Reviewed-by: mmenke <mmenke@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Yao Xiao <yaoxia@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1423164}
NOKEYCHECK=True
GitOrigin-RevId: 1f8da1760171f93036dc6eb5592380f415e25f86