Invalidation tracing: Null-check invalidation sets that might be missing
On carfax.com I was able to reproduce a crash during stylesheet revisit
due to not finding a universal sibling invalidation set that we expected
to find. This crash did not reproduce consistently, which suggests it's
a timing issue similar to https://crbug.com/379170483 where, at the time
tracing starts, a new rule has been added but is not yet indexed into
RuleInvalidationData. When this occurs, rather than mapping the selector
to invalidation set(s) during the revisit phase, we instead map the
selector at the time the relevant invalidation set(s) are subsequently
constructed or added to.

To prevent crashes from occurring in this situation, this CL adds null
checks to spots where we can't assume an invalidation set exists during
the revisit phase.

Fixed: 394523697, 396817121
Bug: 337076014
Change-Id: I4f407c1ece4eb5ad13302436818b4b8893215cb1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6284226
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Kevin Babbitt <kbabbitt@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1423205}
NOKEYCHECK=True
GitOrigin-RevId: b64727240ed6ea6e7c96ff7490ca62d70687f302