GetSourceImageForCanvas: Use kDoNotChangeAlpha in VideoFrame
Change the use of GetSourceImageForCanvas in VideoFrame::Create to not
request an eager conversion to premultiplied alpha.

media::VideoFrame does not track whether or not the alpha channel is
premultiplied. It is implicitly unpremultiplied for YUVA formats. For
RGBA formats, it seems to be implicitly premultiplied.

If the source's image is accelerated and unpremultiplied (which never
happens today, but might one day be enabled), fall back to the software
path.

In the software path, perform an eager conversion to unpremultiplied
format. This preserves the existing behavior prior to this change (and
does not address the question of media::VideoFrame tracking of alpha
type).

Bug: 40277213
Change-Id: Icf2d421eb5d364c524bfdfc97dcb4a6b3503acf3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6281097
Reviewed-by: Colin Blundell <blundell@chromium.org>
Commit-Queue: ccameron chromium <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1423212}
NOKEYCHECK=True
GitOrigin-RevId: 5d698dd55214eaf6d396527f9f23c03009b0ee9c