Add CSS Color 4 gradient options to CanvasGradient
Add ColorInterpolationMethod, HueInterpolationMethod and
premultipliedAlpha attributes to CanvasGradient as described here:

https://github.com/whatwg/html/issues/7947

This is done behind the flag CanvasGradientCSSColor4, as spec is not yet
settled:

https://github.com/whatwg/html/pull/10986

Chromestatus: https://chromestatus.com/feature/5168382242193408

I2P thread:
https://groups.google.com/a/chromium.org/g/blink-dev/c/6PhsIK2kxeM

Gradient::ColorInterpolation is also renamed to
Gradient::PremultipliedAlpha, because that's what it is.

Subpixel antialiasing for test labels is being inconsistently applied on
windows, so the tests are marked as failing with a bug.
crbug.com/396344375

Bug: 394052224
Change-Id: I6b1092376d08c051b161f83c82d585bd03cd4131
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6226021
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: Aaron Krajeski <aaronhk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1423255}
NOKEYCHECK=True
GitOrigin-RevId: 1787a96db896f1973cdbc7a1464137414dbffc68