Add proxy_chain to blink::mojom::ResourceLoadInfo.
This change is necessary to enable User Bypass for IP Protection.

Context: The browser process requires knowledge about if a request is
being proxied for IP Protection. This information can be found in the
`ProxyChain` and is already available in the `URLResponseHead`. This CL
adds the proxy chain to the `ResourceLoadInfo` mojom struct so it can be
accessed there, and populates it in ResourceLoadInfoNotifierWrapper.

Bug: 370696608
Change-Id: I15fe09e66c89adc9247f9e88429956c8149b489e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6266183
Commit-Queue: Abhi Patel <abhipatel@chromium.org>
Reviewed-by: Andrew Williams <awillia@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1423263}
NOKEYCHECK=True
GitOrigin-RevId: 692a428731994e7183b989c665d2652ee7af2f5b