Add map of AXIDs requiring cache update
The change adds a map from AXIDs that require cache update to the
update reason. This map, called `nodes_requiring_cache_update_`, lives
on the AXObjectCacheImpl. AXIDs are added when their corresponding
object's cached attributes need updating, and they are removed when
the object has been removed from the tree, or if their cached
attributes have been updated.

The change adds a setter for `cached_values_need_update_` to AXObject
to allow us to track when a Node needs to be added/removed from the
new map.

This change also moves TreeUpdateReason to a common enums file
so that it can be easily referenced outside of AXObjectCacheImpl.

There were some call sites that added an AXID to
`nodes_requiring_cache_update_` but did not have a corresponding
TreeUpdateReason. To handle these, three new reasons have been added:
kChildInserted, kMarkDocumentDirty, and kNewRelationTargetDirty.

This map is used to confirm there are no objects requiring cached
attribute update after the tree is finalized. We know there are still
some cases where this is happening in the wild, and this change is
meant to help in investigating the root cause of currently possible
and future cases where this occurs. If this map is not empty after
the tree is finalized, DUMP_WILL_BE_CHECK is fired with info on all
remaining entries in the map.

To avoid perf impact, this logic was handled behind AX_FAIL_FAST_BUILD.

Bug: 382104886
Change-Id: I4e9ec8eda7592cadecc522d81f97cb579aa6b806
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6285955
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1423370}
NOKEYCHECK=True
GitOrigin-RevId: eaf2c23ef1fb5cadd47274f5ea16dddc31d9fab1