Handle overlay scrollbars gracefully in auto-scrollbars.html
The WPT test auto-scrollbars.html depends on the browser not having
overlay scrollbars. This test has been consistently failing on mac due
to failing this assertion:
https://luci-milo.appspot.com/ui/tests/p/chromium/rules/751e05390273c6cfc8bf07cc5d245522?annotated=false&historyTimeRange=30d&selectedMetrics=human-cls-failed-presubmit%2Ccritical-failures-exonerated%2Cfailures%2Cbuilds-failed-due-to-flaky-tests&tab=recent-failures&groupBy=os

This CL changes the file so that tests are skipped when overlay
scrollbars are detected instead of crashing after failing the
assertion.

Crbug 398159676 was created as a follow up of this CL to make our
infra handle assert_implements_optional as not failures. After that
work is completed, this CL will be reverted and the assert will be
restored.

Bug: 398159676, 398159680
Change-Id: Ibe6efb7dce6528f6eb397a116ab45312e9edd5cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6288285
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Gaston Rodriguez <gastonr@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1423390}
NOKEYCHECK=True
GitOrigin-RevId: 5a0bc06298ebd68a001ef3d3dba5993de3e3f77b