[css] Fix a an unfortunate typo
This used to be a DCHECK_EQ, but then
https://chromium-review.googlesource.com/c/chromium/src/+/904086/5/third_party/WebKit/Source/core/css/cssom/CSSSkewX.cpp
turned it into an if, without changing the `, ` to a ` == `.

`a, b` evaluates to `b`, so this `if` was always taken.

After https://chromium-review.googlesource.com/c/chromium/src/+/5981211
the NOTREACHED() at the end of the function always crashes,
so it looks like this code should be called only if this `if`
is taken (...and in practice, passing more than one thing to
`transform: skewY()` seems to indeed not crash the renderer).
So maybe this should be a DCHECK() instead of an if after all?

Bug: 393924694
Change-Id: I45896c95ed1fad388a7414910cb20559b8bcba9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6299569
Commit-Queue: Nico Weber <thakis@chromium.org>
Reviewed-by: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1424022}
NOKEYCHECK=True
GitOrigin-RevId: 2bb72000c7ba383b312ababac790216ca229bb81