Map WebDXFeature use counters for Fugu
This CL maps existing counters that are representative of the feature
usage:
- app-file-handlers: WebDXFeatures::kAppFileHandlers is recorded when
launch files are set in the web_launch_service_impl.cc

- app-protocol-handlers: WebDXFeatures::kAppProtocolHandlers is
recorded if the manifest's protocol handler is not empty. Found in the
manifest_manager.cc

- badging: WebDXFeatures::kBadging is recorded when installed web
applications display an application badge. This feature is used in
navigator_badge.cc when setAppBadge() is called.

- manifest: WebDXFeatures::kManifest is recorded when any manifest
metadata about a site is collected. Aside from the name, icon, and
descriptions, this also includes ways in which the app wishes to
integrate with the device.
More info:
https://github.com/web-platform-dx/web
features/blob/main/features/manifest.yml

- beforeInstallPrompt: WebDXFeatures::kBeforeInstallPrompt is
recorded when an event prompt shows up and is completed for a
site that is installable as a PWA.

Change-Id: I069ac1eb8abc2b71ddb800c44b05cc56f218cdae
Bug: 379953791
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6299834
Commit-Queue: May Siem <msiem@chromium.org>
Reviewed-by: Dibyajyoti Pal <dibyapal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1424277}
NOKEYCHECK=True
GitOrigin-RevId: 004786a4bca997d71618c8fc0b3f0c98ea660c27