[functions] Implement tree-scoped lookups
Function names are tree-scoped names [1], and looking up the correct
@function rule for a given <dashed-function> therefore requires the
context of the originating tree scope. This is the first time a
substitution function [2] triggers this behavior, which is why
we now need to plumb the tree scope our numerous Resolve* functions.

This CL leaves some of the tree-scoping work for other CLs:

 - Container lookup uses the wrong tree scope. (CL:6298108)
 - Cycle detection must be tree-scope aware. (CL:6287376)
 - Animations machinery must provide a tree-scope. (CL:6298190)
 - Tests for the ResolveRevert path (blocked on Issue 398554840).
 - Tests for ResolveRevertLayer (blocked on Issue 324780202).

[1] https://drafts.csswg.org/css-scoping-1/#css-tree-scoped-name
[2] https://drafts.csswg.org/css-values-5/#arbitrary-substitution

Bug: 325504770, 394111301
Binary-Size: Not actionable
Change-Id: I6c86d7cc5596f303cb950b49fc3928d0531fe6df
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6297909
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1424529}
NOKEYCHECK=True
GitOrigin-RevId: fbe6362729ef3c06d2e09270a170d48b60791938