scroll anchor: Use priority candidate as root for selection
When an element is a priority candidate, then this
patch uses the element as a root for regular candidate
selection. This ensures that for large contenteditable content, we
still anchor to some child element instead of the whole editable box.

This is also consistent with the type of anchoring we would do
without focus.

R=flackr@chromium.org

Bug: 395489580
Change-Id: Id88477fc07cf0fbb82434e448a9f547314c3db6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6288047
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Vladimir Levin <vmpstr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1424549}
NOKEYCHECK=True
GitOrigin-RevId: aa49c8a3cfa2cee1461281c2552e0899834a3236