[Fontations] Flag to switch system fonts, Chrome UI font to Fontations
Introduce switch and configure system font instantiation with
a FontScanner to ensure that instantiation happens through
Fontations instead of FreeType.

The outcome of this change has been tested in [1]: If this flag is
switched on, Linux system font in content and in the Chrome browser UI
on Linux and ChromeOS will move to being rendered by Fontations. This
can mean tiny pixel changes in antialiasing, but from a user experience
point of view, this is intended as an unnoticeable change.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/6169919

Bug: chromium:346918516
Change-Id: Id4ae3094083c9b964854c5e7cef31dc6c6907e29
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6286047
Reviewed-by: Ben Wagner <bungeman@google.com>
Auto-Submit: Dominik Röttsches <drott@chromium.org>
Commit-Queue: Ben Wagner <bungeman@google.com>
Cr-Commit-Position: refs/heads/main@{#1424559}
NOKEYCHECK=True
GitOrigin-RevId: 493caf92f7f952182c4907472ba462e5742bd5e0