Move Path to platform/geometry
Path/FloatRoundedRect are co-dependent and should live in the same
place. "Path" is (arguably) more of a geometry primitive than
FloatRoundedRect is a graphics primitive, so this CL moves Path
into platform/geometry. It pulls a few stroke-related primitives and
utils with it.

Bug: 393145930

R=fs@opera.com
R=pdr@chromium.org

Change-Id: I1716e17d53388a8e7eebfe2af88ee8c1dbea9e93
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6282150
Commit-Queue: Noam Rosenthal <nrosenthal@chromium.org>
Reviewed-by: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/main@{#1424616}
NOKEYCHECK=True
GitOrigin-RevId: 0ddcd55dea33f343ce1b78fce47970af273d67e0