Cache the resolved versions of packages in the androidx subproject
Now that we are using '+' to roll androidx packages to latest snapshot,
gradle needs to re-evaluate what the '+' maps to which takes a long
time. Now we cache the resolved versions so we do not need to resolve
the '+' unless we are trying to roll to a new snapshot.

Change-Id: I4d0b3a52a9ab88e7e6ce86738d1ffcd24a11a78d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6298139
Commit-Queue: Peter Wen <wnwen@chromium.org>
Auto-Submit: Mohamed Heikal <mheikal@chromium.org>
Commit-Queue: Mohamed Heikal <mheikal@chromium.org>
Reviewed-by: Peter Wen <wnwen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1424709}
NOKEYCHECK=True
GitOrigin-RevId: 811bb64916d3f8e04ca54212edf70d6b8b72373f