Prompt API: add sampling param checking logic from availability()
This CL checks if the sampling param passed in the `availability()` is
valid using the `ResolveSamplingParamsOption()` method. The sampling
params will be checked from the renderer and the mojom interface is
modified so only the `expected_input_languages` will be passed to the
browser.

Change-Id: I17fb73624a13f2863cd4869bba3071ff1251f8cc
Bug: 390459309
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6277982
Commit-Queue: Mingyu Lei <leimy@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1424931}
NOKEYCHECK=True
GitOrigin-RevId: b5b6a08629d3c945b10047b7e2f169ddfe565c9a