Fix carriage return DCHECK in LineBreaker::SplitTrailingBidiPreservedSpace
In crrev.com/c/5199988, the method `SplitTrailingBidiPreservedSpace`
was added to the `LineBreaker` class. When called on a line with
trailing spaces, and which has bidi enabled, it will iterate through
the line's text or control items in reverse. And at one point there's
a DCHECK asserting that the item's end offset must be greater than
zero.

This would fail for any empty items, but text or control items tend
not to be empty. However, one thing that creates an empty control item
is a 0x0D carriage return character in the DOM text. (This does not
happen with a carriage return in the HTML source, since it gets
converted into a 0x0A newline when parsing; it needs a character
reference in the source, or insertion from JS.)

This patch fixes this by detecting whether a text or control item is
empty before that DCHECK, and treating it as part of the trailing
spaces if it is.

Bug: 398527874
Change-Id: Ibb9bf842aeb9ab3614d8b9c9d53f15b3d18557e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6298873
Commit-Queue: Andreu Botella <abotella@igalia.com>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1424977}
NOKEYCHECK=True
GitOrigin-RevId: cfba7a63d8273e6bba67f0c749b7c20cbb5dddb4