Clear measure cache when rebuilding fragment tree spine.
When adding layout results (and their fragments) in the first place, and
the layout result is a measure-result, we'll add the result *both* to
the measure cache and to the vector of regular layout results. Because
of this, when rebuilding the fragment tree spine after sub-tree layout
of a relayout boundary object, remove any measure results, since they
are potentially outdated.

Bug: 395709683
Change-Id: Idba4fe74d3f8fb9a416d248211a096da0ef1ed66
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6298020
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Reviewed-by: David Grogan <dgrogan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1425015}
NOKEYCHECK=True
GitOrigin-RevId: a47c0ac75fdf5dd748392ed73906c98674fb33ca