Remove new_provider_for_testing argument
This parameter was only used by a single unit test:
DisableAccelerationPreservesRasterAndRecording. This unit test used a
fake provider to check that the a flush is performed and a raster is
written when disabling GPU acceleration. This behavior is already
covered by the following web test, so we don't need to validate it as
a unit test too:
fast/canvas/disable-acceleration-preserves-recording.html

Bug: 346766781
Change-Id: I98c423a6cf01b7c91ff17e8c6d89fe4c6ac57b2b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6305963
Reviewed-by: Colin Blundell <blundell@chromium.org>
Commit-Queue: Jean-Philippe Gravel <jpgravel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1425217}
NOKEYCHECK=True
GitOrigin-RevId: 99273838554a75e60a79a358070cff4ff3ab4fb1