cv/carousel: Always style c-v auto locks under scroll-marker-group
This patch tracks when content visibility lock is under
an ancestor wiwth a scroll-marker-group. If that's the case, and the
lock is auto, then it permits style. It also permits layout for CQ.

It also tracks when c-v is added first and then scroll-marker-group and
vice versa. It also tracks when the lock is moved to a different
document. This is all similar to, for example, focus tracking.

R=chrishtr@chromium.org, futhark@chromium.org, flackr@chromium.org

Bug: 395078320
Change-Id: I8d89ee0bc3104f572c762f1d38880ad0e19bde23
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6241701
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Vladimir Levin <vmpstr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1425264}
NOKEYCHECK=True
GitOrigin-RevId: 7aa0764dfdb0c15d8559717b85f54134d7ddf16b