[PWA] Add WebFeature usage counter for manifest update
This CL adds a WebFeature use counter for manifest updating and ties
it to the WebDXFeature for kManifest, so that manifest updating also
start measuring usage of the manifest feature. This is done by hooking
up the MetricsObserver on the browser side.

Tested by verifying 2 use-cases, a successful manifest update measures
this feature, while an unsuccessful one doesn't.

Bug: 399128157, 379953791
Change-Id: I5ab57cf887885e05273a3f038501b64aa1524469
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6303004
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Reviewed-by: May Siem <msiem@chromium.org>
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1425316}
NOKEYCHECK=True
GitOrigin-RevId: ad07ffe6771c2897dd8b99cb2fb726617cf56d9f