Remove alph@ from //third_party/inspector_protocol/OWNERS.
By policy, all owners must be committers. alph@chromium.org
has been inactive on the project for some time and is no longer
a committer as a result, so this CL removes him from the file
where he was listed as an OWNER.

(See
https://chromium.googlesource.com/chromium/src/+/HEAD/docs/code_reviews.md#expectations-of-owners
for the policy).

Bug: 376077658
Change-Id: Ibc7e2c628526552ad248ce161529ad40babf6cd5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6279051
Commit-Queue: Dirk Pranke <dpranke@google.com>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1425348}
NOKEYCHECK=True
GitOrigin-RevId: 25a3c804380ede30131c6ecd9f646473c30c04eb