[Grid] Oilpan-ify Grid code
This change updates many of the internal grid data structures to be GC'd
via Oilpan.

In GridSizingTree::GridTreeNode, I moved a public struct field to a
getter/setter to ensure that Member-ified `grid_items` is always
available, instead of a multitude of nullptr checks. This also required
the Member to be mutable to maintain const-ness.

There are some expected perf costs to using Oilpan - up to 20% in
synthetic tests, but there aren't any statsig differences in
Speedometer, so actual users shouldn't see a difference in performance.
We have discussed some ideas for mitigating this in the future, tracked
by crbug.com/399153019.

Bug: 394634504

Change-Id: I8ebd48a67961d4a45341b8dd2a26c99854b26407
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6278548
Commit-Queue: Kurt Catti-Schmidt <kschmi@microsoft.com>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Reviewed-by: Ethan Jimenez <ethavar@microsoft.com>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1425546}
NOKEYCHECK=True
GitOrigin-RevId: d25c718dde23768e13169395c4cc52d5bdb4bd47