WebGL transferToImageBitmap: Add diagnostic message to exception
If we we see that the context is lost (because GetDrawingBuffer fails),
then include this information in the exception thrown by
transferToImageBitmap.

This will not solve the problem in the bug, but may help in diagnosing
the problem (and is generally more useful than the more generic
exception messages).

Bug: 397998415
Change-Id: I8d76dd418e44ffba2e95922c88151f2d052f34b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6308772
Reviewed-by: Colin Blundell <blundell@chromium.org>
Commit-Queue: ccameron chromium <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1426270}
NOKEYCHECK=True
GitOrigin-RevId: 5aba9f86fe7cb0aa641b92af535b461814995665