[FONT] Addresses Font fallback failure for Nirmala UI
The issue is that none of the glyphs for the Meitei Mayek
script are rendering correctly with the Nirmala UI font.
All glyphs appear as tofu because the ScriptToFontFamilies
map is missing the key-value pair for Meitei Mayek to the
corresponding font family(Nirmala UI).

Bug: 350029227
Change-Id: I823381ad6dc68dbf6aceb35ddb435be3314f16da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6303758
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Dileep Maurya <dileepmaurya@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1426335}
NOKEYCHECK=True
GitOrigin-RevId: 12e618e25d35bee687ab0202bc91bbde95da86e8