[Blink] Eliminate need for clients to check for kSharedBitmap CRP type
The default impl of
CanvasResourceProvider::IsSharedBitmapGpuChannelLost() returns false,
with CanvasResourceProviderSharedImage overriding it to check whether
the SharedBitmap gpu channel (now used for software SharedImages) is
lost. By having the latter return false if `is_software_` is false, we
can eliminate the need for clients to check for the type of the
CanvasResourceProvider being `kSharedBitmap` before calling this.

This will enable deprecating the `kSharedBitmap` type in a followup.

Bug: 379996128
Change-Id: I4439173939e5582086e5eaf144b12544ebee2d21
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6308729
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1427040}
NOKEYCHECK=True
GitOrigin-RevId: 977d99cfeeaeb7c0eaa3d701388b3aad1d381648